refactor: separate Redux thunks from services in Recents module
