+++ src/views/Recents/index.ts
@@ -1,3 +1,3 @@
 export { default } from './RecentsView';
-export { fetchRecents } from './services/fetchRecents';
-export { fetchRecentsThunk, fetchRecentsThunkExtraReducers } from './services/fetchRecentsThunk';
+export * from './services';
+export * from './store';
