+++ src/views/Recents/services/index.ts
@@ -1,0 +1,1 @@
+export { fetchRecents } from './fetchRecents';
