+++ src/views/Recents/store/index.ts
@@ -1,0 +1,1 @@
+export { fetchRecentsThunk, fetchRecentsThunkExtraReducers } from './fetchRecentsThunk';
